--- a/.npmrc
+++ b/.npmrc
@@ -1,5 +1,6 @@
 # pnpm workspace configuration
-hoist=false
+# Use shamefully-hoist for better performance in monorepos
+shamefully-hoist=true
 
 # Auto install peer dependencies
 auto-install-peers=true
@@ -7,8 +8,8 @@ auto-install-peers=true
 # Strict peer dependencies
 strict-peer-dependencies=false
 
-# Use hard links for better performance
-package-import-method=hardlink
+# Use copy for better compatibility and performance
+package-import-method=copy
 
 # Link workspace packages by default
 link-workspace-packages=true
@@ -16,10 +17,23 @@ link-workspace-packages=true
 # Save workspace protocol
 save-workspace-protocol=rolling
 
-# Recursive install
-recursive-install=true
-
 # Enable lockfile validation
 prefer-frozen-lockfile=true
-public-hoist-pattern[]=@apollo/*
-public-hoist-pattern[]=@types/* 
+
+# Store directory for better performance
+store-dir=~/.pnpm-store
+
+# Side effects cache for faster installs
+side-effects-cache=true
+
+# Verify store integrity
+verify-store-integrity=false
+
+# Network settings for faster downloads
+network-concurrency=16
+fetch-retries=3
+fetch-retry-factor=2
+fetch-retry-mintimeout=10000
+fetch-retry-maxtimeout=60000
+
+