--- a/client/src/pages/CardsNav.tsx
+++ b/client/src/pages/CardsNav.tsx
@@ -1,8 +1,6 @@
 import React from 'react';
-import { useStyles } from '../styles';
+import { NavRoot } from '../styles';
 
 export const CardsNav: React.FC = (): React.ReactElement => {
-  const classes = useStyles();
-
-  return <div className={classes.navRoot}>WIP</div>;
+  return <NavRoot>WIP</NavRoot>;
 };