--- a/package.json
+++ b/package.json
@@ -1,7 +1,7 @@
 {
   "name": "brainstrike-monorepo",
   "private": true,
-  "packageManager": "pnpm@9.6.0",
+  "packageManager": "pnpm@10.11.1",
   "engines": {
     "node": ">=18.19.1",
     "pnpm": ">=9.6.0"
@@ -25,24 +25,24 @@
     "typecheck": "pnpm --recursive typecheck"
   },
   "devDependencies": {
-    "@graphql-codegen/cli": "5.0.5",
-    "@graphql-codegen/typescript": "4.1.5",
-    "@types/node": "22.13.8",
-    "@typescript-eslint/eslint-plugin": "8.25.0",
-    "@typescript-eslint/parser": "8.25.0",
-    "eslint": "9.21.0",
-    "eslint-config-prettier": "10.0.2",
-    "eslint-plugin-prettier": "5.2.3",
-    "prettier": "3.5.2",
-    "syncpack": "^12.3.0",
-    "typescript": "5.8.2"
+    "@graphql-codegen/cli": "5.0.7",
+    "@graphql-codegen/typescript": "4.1.6",
+    "@types/node": "22.15.30",
+    "@typescript-eslint/eslint-plugin": "8.33.1",
+    "@typescript-eslint/parser": "8.33.1",
+    "eslint": "9.28.0",
+    "eslint-config-prettier": "10.1.5",
+    "eslint-plugin-prettier": "5.4.1",
+    "prettier": "3.5.3",
+    "syncpack": "^13.0.4",
+    "typescript": "5.8.3"
   },
   "pnpm": {
     "overrides": {
-      "graphql": "16.10.0",
+      "graphql": "16.11.0",
       "@types/ws": "8.5.4",
-      "typescript": "5.8.2",
-      "eslint": "9.21.0"
+      "typescript": "5.8.3",
+      "eslint": "9.28.0"
     }
   }
 }