--- a/server/package.json
+++ b/server/package.json
@@ -60,7 +60,7 @@
     "apollo-link-http": "1.5.17",
     "eslint": "9.28.0",
     "eslint-config-prettier": "10.1.5",
-    "eslint-import-resolver-typescript": "3.8.3",
+    "eslint-import-resolver-typescript": "4.4.3",
     "eslint-plugin-import": "2.31.0",
     "eslint-plugin-prettier": "5.4.1",
     "nock": "14.0.5",