… logo.png
… pnpm-lock.yaml
… pnpm-workspace.yaml
… screenshot-client.png