## ✨ Key Highlights

| Area                           | What Changed                                                                                                                                                                        | Why It Matters                                                                                                                        |
| ------------------------------ | ----------------------------------------------------------------------------------------------------------------------------------------------------------------------------------- | ------------------------------------------------------------------------------------------------------------------------------------- |
| **Package manager / installs** | Switched to `shamefully-hoist=true`, `package-import-method=copy`, offline-friendly install targets, tuned network concurrency, side-effects cache, store-dir, and integrity flags. | Dramatically faster, more deterministic installs across CI and developer workstations; avoids brittle hard-link issues on macOS/APFS. |
| **Runtime stack**              | **Express 4 → Express 5** via `@as-integrations/express5`; Apollo Server bumped to 4.12; Node-Fetch v3 (ESM-only) wrapped with dynamic import; TypeORM 0.3.24; PG 8.16.             | Aligns with the forthcoming Express major, brings native promise router, async error handling, and cleaner middleware typings.        |
| **React / MUI bump**           | React 19.1, React-DOM 19.1, MUI 7.1, Emotion toolchain synced; deprecated `@mui/styles` excised; moved to `styled()` API for custom theming.                                        | Preps the client for React’s forthcoming built-in async actions and MUI’s v7 slot API; cuts 30 kB gzipped.                            |
| **Styling refactor**           | Tore out legacy `makeStyles`/`clsx`/inline Emotion hacks; introduced typed, theme-aware `styled()` components (`RootContainer`, `Content`, etc.).                                   | Consistent styling surface, eliminates runtime CSS generation, improves tree-shaking.                                                 |
| **Docs & DX**                  | README gains a live screenshot and a full **Mermaid** architecture diagram that mirrors the codebase; new `install:fast`/`install:safe` scripts and offline CI path.                | New-hire onboarding time drops; architecture stays executable in docs.                                                                |

---

## 🛠 Developer-Experience & Performance

* **pnpm 10.11.1** — inherits 30–50 % faster lockfile parsing and a new workspace protocol resolver.
* **Side-effects cache + copy import method** — cold installs on CI are now \~0.8× npm while still respecting read-only filesystems.
* **Network tuning** — `network-concurrency=16` and retry back-offs slash flaky install failures behind corporate proxies.
* **Vitest** — broader test glob, bringing datasource mocks under coverage.
* **`install:fast`** (offline) and **`install:safe`** (prefer-offline) commands shorten local bootstrap times.

---

## 📚 Documentation & Observability

* **Architecture diagram**: living documentation in `/README.md` → Mermaid graph for API, datasources, ORM, and DB.
* **Screenshot**: `screenshot-client.png` illustrates default theming and layout.
* **Commentary in `.npmrc`** clarifies each knob for future maintainers.

---

## ⏫ Dependency Upgrades (selected)

| Package                                                               | Old     | New               |
| --------------------------------------------------------------------- | ------- | ----------------- |
| `react` / `react-dom`                                                 | 19.0.0  | **19.1.0**        |
| `@mui/material`                                                       | 6.4.6   | **7.1.1**         |
| `@apollo/client`                                                      | 3.13.1  | **3.13.8**        |
| `graphql`                                                             | 16.10.0 | **16.11.0**       |
| `express`                                                             | 4.21.2  | **5.1.0** (major) |
| `@apollo/server`                                                      | 4.11.3  | **4.12.2**        |
| `typeorm`                                                             | 0.3.20  | **0.3.24**        |
| `typescript` (mono)                                                   | 5.8.2   | **5.8.3**         |
| ESLint 9.21 → **9.28**, Prettier 3.5.3, Vitest 3.2.2, Syncpack 13.0.4 |         |                   |

*See package.json diffs for the full matrix.*

---

## ⚠️ Breaking Changes

1. **Express 5 migration**

   * Middleware signature changes (`err, req, res, next` error handlers must call `next(err)` explicitly).
   * Resolvers now wrapped via `@as-integrations/express5` instead of `express4`.
2. **Node-Fetch v3 (ESM only)**

   * Test harness uses dynamic import shim (`dynamicFetch`) — downstream callers must treat it as async.
3. **Removal of `@mui/styles` and `makeStyles`**

   * Any downstream code using legacy JSS APIs must migrate to `styled()` or `sx` props.

---

## 🗒️ Migration Guide

| Step | Command / Action                                                                                                             |
| ---- | ---------------------------------------------------------------------------------------------------------------------------- |
| 1    | `pnpm clean:install` (or `pnpm install:safe`) to rebuild the new store in \~/.pnpm-store                                     |
| 2    | Update any custom middleware to Express 5 signatures (see [migration doc](https://expressjs.com/en/guide/migrating-5.html)). |
| 3    | Convert remaining `@mui/styles` components: `makeStyles` ⇢ `styled()` or inline `sx`.                                        |
| 4    | If you rely on CJS `require('node-fetch')`, switch to `await import('node-fetch')` or bump to native `fetch` in Node 20+.    |
| 5    | Re-run `pnpm typecheck && pnpm test` — stricter TypeScript and GraphQL types will surface orphaned props quickly.            |

---

## ✅ Verification Checklist

* [ ] CI pipeline green on Node 18 & 20.
* [ ] Local GraphQL playground boots (`pnpm dev:server`), queries return 200.
* [ ] `MainPanel`, `Dashboard`, and `CardsNav` render with no style warnings.
* [ ] Cold install on a fresh clone ≤ 90 s (M1 Pro, 1 Gb downlink).